completed backend

+++ backend/controller/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 //@desc Get Goals
 //@route /api/goals
 //@access private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 //@desc Set Goal
@@ -17,6 +18,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -29,6 +31,21 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("No Goal Found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Checking if logged in user and the goal id matches
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not Authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -43,6 +60,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("No Goal Found");
   }
+  const user = await User.findById(req.user.id);
+
+  //Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Checking if logged in user and the goal id matches
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not Authorized");
+  }
+
   await goal.deleteOne();
   res.status(200).json({ id: req.params.id });
 });
